==> data_preparation/00_addresses.ipynb <==
import requests
from bs4 import BeautifulSoup
from tqdm.auto import tqdm
import pandas as pd

page = requests.get(url='https://xn--d1abbjabeif1aexoaci4o.xn--p1ai/postcodes/moskva/0/moskva/')
page = BeautifulSoup(page.text)

letters = []
for link in page.find_all('a', {'class': 'link'}):
    letters.append(link.attrs['href'].split('=')[-1])

link.attrs['href']

print(letters)

letters = ['1', 'a', 'b', 'v', 'g', 'd', 'e', 'zh', 'z', 'i', 'k', 'l', 'm', 'n', 'o', 'p', 'r', 's', 't', 'u', 'f', 'h', 'ts', 'ch', 'sh', 'shch', 'e', 'yu', 'ya']

full_list = []

for letter in tqdm(letters):
    page = requests.get(
        url=f'https://xn--d1abbjabeif1aexoaci4o.xn--p1ai/postcodes/moskva/0/moskva/?letter={letter}'
    )
    page = BeautifulSoup(page.text)  
    for element in page.find('div', {'class': 'list_index'}).find_all('div', {'class': 'text-left'}):
        full_list.append((element.find('span').get_text(), element.find('a').get_text()))

index_data = pd.DataFrame(full_list, columns=['postal_code', 'street'])
index_data = index_data[index_data['postal_code'] != '']
index_data['postal_code'] = index_data['postal_code'].astype(int)

index_data

index_data.to_csv('postal_codes.csv', index=None)
