==> data_preparation/00_categories.ipynb <==
import pandas as pd

dct = pd.read_excel('../data/dict.xlsx')

dct.head()

dct['level1'] = dct['level1'].apply(lambda x: x.replace('ОНЛАЙН', '').strip() if 'ОНЛАЙН' in x else x)
dct['level2'] = dct['level2'].apply(lambda x: x.replace('ОНЛАЙН', '').strip() if 'ОНЛАЙН' in x else x)
dct['leve3'] = dct['leve3'].apply(lambda x: x.replace('ОНЛАЙН', '').strip() if 'ОНЛАЙН' in x else x)

dct = dct.groupby(['level1', 'level2', 'leve3'], as_index=False).agg({
    'Разметка: Для ума/ Для души / Для тела': 'first',
    'id_level1': 'min',
    'id_level2': 'min',
    'id_level3': 'min',
})

dct = dct[['id_level1', 'id_level2', 'id_level3', 'level1', 'level2', 'leve3', 'Разметка: Для ума/ Для души / Для тела']]
dct.columns = ['cat1_id', 'cat2_id', 'cat3_id', 'cat1_name', 'cat2_name', 'cat3_name', 'branch']

dct.to_csv('../data/dict.csv')
