==> data_preparation/03_model_categories.ipynb <==
import sqlite3

import pandas as pd
import numpy as np
from tqdm.auto import tqdm
import seaborn as sns
import matplotlib.pyplot as plt

db = sqlite3.connect('../raw_db.db')

dct = pd.read_csv('../data/dict.csv').drop('Unnamed: 0', axis=1)

dct.columns = [
        'category1_id'
        , 'category2_id'
        , 'category3_id'
        , 'category1_name'
        , 'category2_name'
        , 'category3_name'
        , 'category'
]

dct.head()

## Предсказание категорий

group_user_stats = pd.read_sql("""
    SELECT 
        category3_id as category_in
        , a.user_id as user_id
        , count(event_id) as n
        , max(is_woman) as is_woman
        , CAST((julianday('2023-03-01') - julianday(date_of_birth)) / 365 / 10 AS INT) * 10 as age_group
    FROM attend a
    INNER JOIN users u ON u.user_id = a.user_id
    INNER JOIN groups g ON a.group_id = g.group_id
    WHERE event_date < '2023-02-01'
        AND a.user_id % 5 = 0
    GROUP BY category_in, a.user_id
""", con=db).dropna().astype(int)

# group_info = pd.read_sql("""
#     SELECT group_id, category3_id as category_in
#     FROM groups
# """, con=db).dropna().astype(int) # (0)

# group_user_stats = group_user_stats.merge(group_info)

group_user_stats

age_rank = pd.read_sql("SELECT category3_id as category_out, is_woman, age_group, rank as rank_age FROM model_age_category", con=db).dropna()
age_rank.head()

category_rank = pd.read_sql("""
SELECT i as category_in, j as category_out, rank as rank_cat 
FROM model_category_similarity
""", con=db).dropna()
category_rank.head()

category_rank = pd.concat([
    category_rank, 
    pd.DataFrame([(i, i, 0) for i in dct['category3_id']], columns=category_rank.columns)
])

n_visits = group_user_stats.groupby('user_id', as_index=False).agg({'n': 'sum'})
n_visits.columns = ['user_id', 'user_visit_total']

group_user_stats = group_user_stats.merge(category_rank)

group_user_stats['is_same'] = (group_user_stats['category_out'] == group_user_stats['category_in']) * group_user_stats['n']

group_user_stats['rank_weight'] = group_user_stats['rank_cat'] * group_user_stats['n']

group_user_stats_grouped = group_user_stats.groupby(['user_id', 'category_out'], as_index=False).agg({
    'age_group': 'first',
    'is_woman': 'first',
    'n': 'sum',
    'rank_cat': 'sum',
    'category_in': 'count',
    'is_same': 'max',
})

group_user_stats_grouped = group_user_stats_grouped.merge(n_visits)
group_user_stats_grouped.head()

group_user_stats_grouped['norm_rank_cat'] = (-1 * group_user_stats_grouped['rank_cat'] / group_user_stats_grouped['n']) / group_user_stats_grouped['user_visit_total']

group_user_stats_grouped = group_user_stats_grouped.merge(
    group_user_stats_grouped[['user_id', 'age_group', 'is_woman']].drop_duplicates().merge(age_rank)
    , how='outer'
)

group_user_stats_grouped = group_user_stats_grouped[['user_id', 'category_out', 'is_same', 'norm_rank_cat', 'rank_age']]

group_user_stats_grouped = group_user_stats_grouped.fillna({'is_same': 0, 'norm_rank_cat': 50})

group_user_stats_grouped

group_user_stats_target = pd.read_sql("""
    SELECT 
        category3_id as category_out
        , a.user_id as user_id
        , count(a.user_id) as n_occurred
        , 1 as has_occurred
    FROM attend a
    INNER JOIN users u ON u.user_id = a.user_id
    INNER JOIN groups g ON a.group_id = g.group_id
    WHERE event_date >= '2023-02-01'
        AND a.user_id % 5 = 0
    GROUP BY category_out, a.user_id
""", con=db).dropna().astype(int)

group_user_stats_data = group_user_stats_grouped.merge(group_user_stats_target, on=['user_id', 'category_out'], how='left')

group_user_stats_data = group_user_stats_data[group_user_stats_data['user_id'].isin(group_user_stats_target['user_id'].unique())]

group_user_stats_data = group_user_stats_data.fillna(0)

group_user_stats_data.head(10)

from sklearn.model_selection import train_test_split

users_train, users_test = train_test_split(group_user_stats_data['user_id'].unique(), test_size=0.2)

i_train = group_user_stats_data['user_id'].isin(set(users_train))
i_test = group_user_stats_data['user_id'].isin(set(users_test))

X_train = group_user_stats_data[i_train].set_index(['user_id', 'category_out'])[['norm_rank_cat', 'rank_age', 'is_same']]
X_test = group_user_stats_data[i_test].set_index(['user_id', 'category_out'])[['norm_rank_cat', 'rank_age', 'is_same']]
y_train = group_user_stats_data[i_train].set_index(['user_id', 'category_out'])['has_occurred']
y_test = group_user_stats_data[i_test].set_index(['user_id', 'category_out'])['has_occurred']

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report

model = LogisticRegression(class_weight='balanced')

model.fit(X_train, y_train)

print(classification_report(y_test, model.predict(X_test)))

x_see = X_test.copy()
x_see['n_occurred'] = group_user_stats_data[group_user_stats_data['user_id'].isin(set(users_test))]['n_occurred'].values
x_see['has_occurred'] = group_user_stats_data[group_user_stats_data['user_id'].isin(set(users_test))]['has_occurred'].values
x_see['predict'] = model.predict_proba(X_test)[:, 1]
x_see = x_see.reset_index()

# sns.lmplot('norm_rank_cat', 'rank_age', data=x_see)

x_see.head()

x_see[['norm_rank_cat', 'rank_age', 'n_occurred']].corr()

x_see = x_see.sort_values(by='predict', ascending=False).groupby('user_id', as_index=False).agg({
    'predict': list,
    'category_out': list,
    'rank_age': list,
    'norm_rank_cat': list,
})

x_see.columns = ['user_id', 'predict', 'category_pred', 'age', 'cat']

# x_see['n_predictions'] = x_see['category_out'].apply(len)
# x_see['ok'] = x_see[['category_out', 'has_occurred']].apply(
#     lambda x: [i for i, v in enumerate(zip(x['category_out'], x['has_occurred'])) if v[1] == 1]
#     , axis=1
# )

group_user_stats_target_all = group_user_stats_target.sort_values(by='n_occurred', ascending=False).groupby('user_id', as_index=False).agg({
    'n_occurred': list,
    'category_out': list,
    'category_out': list
})

x_see = x_see.merge(group_user_stats_target_all)

x_see['share_uq'] = x_see.apply(
    lambda x: np.average([1 if real in x['category_pred'][:10] else 0 for real in x['category_out'][:10]])
    , axis=1
)

x_see['share_uq'].mean()

x_see['share_n'] = x_see.apply(
    lambda x: 
    np.sum([n if real in x['category_pred'][:10] else 0 for real, n in zip(x['category_out'], x['n_occurred'])]) /
    np.sum(x['n_occurred'][:10])
    , axis=1
)
x_see['share_n'].mean()

print(model.intercept_)
for n, c in zip(X_train.columns, model.coef_[0]):
    print(c, n)

from collections import Counter, defaultdict

extra = Counter()
missed = Counter()
correct = Counter()

for i in x_see.index:
    row = x_see.loc[i]
    pred = set(row['category_pred'][:10])
    real = set(row['category_out'])
    correct += Counter(pred & real)
    extra += Counter(pred - real)
    missed += Counter(real - pred)

dct[dct['category3_id'].isin([i[0] for i in extra.most_common(10)])]

correct.most_common(10)

missed.most_common(10)

extra.most_common(10)

x_see

# x_see['ok'].apply(lambda x: sum(1 for i in x if i < 10)).mean()

# x_see['ok'].apply(lambda x: sum(1 for i in x if i < 10)).mean()

# x_see['ok'].apply(lambda x: sum(1 for i in x)).mean()
