==> data_preparation/03_model_groups-Copy1.ipynb <==
# ! pip3 install --upgrade pandas

import sqlite3

import pandas as pd
import numpy as np
from tqdm.auto import tqdm
import seaborn as sns
import matplotlib.pyplot as plt

tqdm.pandas()

db = sqlite3.connect('../raw_db.db')

## Группы в феврале

group_time_target = pd.read_sql("""
    SELECT 
        group_timetable.group_id 
        , weekday
        , CAST(start_time AS INT) as start_time
        , CAST(finish_time AS INT) as finish_time
        , CASE WHEN finish_date > '2023-02-28' AND start_date < '2023-02-01' 
            THEN julianday('2023-02-28') - julianday('2023-02-01')
            WHEN finish_date > '2023-02-28' THEN julianday('2023-02-28') - julianday(start_date)
            WHEN start_date < '2023-02-01' THEN julianday(finish_date) - julianday('2023-02-01')
            ELSE julianday(finish_date) - julianday(start_date)
            END as n_days
    FROM group_timetable
    WHERE start_date < '2023-02-28' AND finish_date > '2023-02-01'
""", con=db, dtype={
    'group_id': 'int32', 'weekday': 'int8', 'start_time': 'int8', 'finish_time': 'int8', 'n_days': 'int8'
})

def week_ones(d, s, f, n=1, padding=0):
    a = np.zeros((7, 17), dtype='int8')
    a[d - 1, max((s - 6 - padding), 0):(f - 6 + padding)] = n
    return a

group_time_target['timetable'] = group_time_target.apply(lambda x: week_ones(x['weekday'], x['start_time'], x['finish_time']), axis=1)

group_time_target = group_time_target.groupby('group_id', as_index=False).agg({'timetable': sum, 'n_days': 'min'})

group_time_target.head()

group_place_target = pd.read_sql("""
    SELECT 
        group_id
        , district_id
        , zone_id
        , COALESCE(code, 0) as g_postal
    FROM group_locations
    WHERE group_id IN (
        SELECT DISTINCT group_id 
        FROM group_timetable
        WHERE start_date < '2023-02-28' AND finish_date > '2022-02-01'
    )
""", con=db, dtype={'group_id': 'int32', 'district_id': 'int8', 'zone_id': 'int8', 'g_postal': 'int32'})

group_place_target = group_place_target.groupby('group_id', as_index=False).agg({
    'district_id': tuple,
    'zone_id': tuple,
    'g_postal': lambda x: tuple(i for i in x if i > 0)
})

group_info = pd.read_sql("""
    SELECT 
        group_id
        , category3_id
        , category2_id
        , is_online
        , is_mobile
    FROM groups
    WHERE group_id IN (
        SELECT DISTINCT group_id 
        FROM group_timetable
        WHERE start_date < '2023-02-28' AND finish_date > '2023-02-01'
    )
""", con=db, dtype={'group_id': 'int32', 'category3_id': 'int16', 'category2_id': 'int16', 'is_online': 'int8', 'is_mobile': 'int8'})

group_target = group_time_target.merge(group_place_target).merge(group_info)

group_target.head()

postat_to_zd = pd.read_sql("""
    SELECT district_id, zone_id, code, count(group_id) as n
    FROM group_locations
    WHERE code IS NOT NULL
    GROUP BY district_id, zone_id, code
    ORDER BY n DESC
""", con=db, dtype={'district_id': 'int16', 'zone_id': 'int16', 'code': 'int32', 'n': 'int16'}).drop_duplicates(subset=['code'], keep='first')

code_dict = {c: (z, d) for d, z, c, _ in postat_to_zd.values}

group_target_real = pd.read_sql("""
    SELECT 
        attend.user_id
        , group_id
        , count(attend.user_id) as n
        , 1 as visited
        , COALESCE(postal_code, 0) as postal_code
    FROM attend
    LEFT JOIN users ON attend.user_id = users.user_id
    WHERE event_date >= '2023-02-01'
    GROUP BY attend.user_id, group_id
""", con=db, dtype={'user_id': 'int32', 'group_id': 'int32', 'n': 'int8', 'visited': 'int8', 'postal_code': 'int32'})

group_target_real = group_target_real.merge(group_target[['category2_id', 'category3_id', 'group_id']])

group_target_real.columns = ['user_id', 'group_id', 'n_visits', 'visited', 'u_postal', 'category2_id', 'category3_id']

group_target_real['u_district'] = group_target_real['u_postal'].apply(lambda x: code_dict.get(x, [0, 0])[1]).astype('int16')
group_target_real['u_zone'] = group_target_real['u_postal'].apply(lambda x: code_dict.get(x, [0, 0])[0]).astype('int16')

user_flat = group_target_real[['user_id', 'category3_id', 'u_postal', 'u_district', 'u_zone']].drop_duplicates()

all_combinations = group_target.merge(user_flat, on=['category3_id'])

geo_sim = all_combinations[['district_id', 'zone_id', 'g_postal', 'u_postal', 'u_district', 'u_zone']].drop_duplicates()

geo_sim['same_post'] = geo_sim[['u_postal', 'g_postal']].progress_apply(lambda x: x['u_postal'] in x['g_postal'], axis=1).astype('int8')
geo_sim['same_zone'] = geo_sim.progress_apply(lambda x: x['u_zone'] in x['zone_id'], axis=1).astype(int).astype('int8')
geo_sim['same_district'] = geo_sim.progress_apply(lambda x: x['u_district'] in x['district_id'], axis=1).astype(int).astype('int8')

# geo_sim['g_postal'].unique()

neighbors = pd.read_sql("""
    SELECT group_id, postal_code, n_neighbors
    FROM model_neighbors
""", con=db, dtype={'group_id': 'int32', 'postal_code': 'int32', 'n_neighbors': 'int16'})

all_combinations = all_combinations.merge(
    geo_sim, how='left'
).merge(
    neighbors, left_on=['group_id', 'u_postal'], right_on=['group_id', 'postal_code'], how='left'
)

all_combinations['n_neighbors'] = all_combinations['n_neighbors'].fillna(0).astype('int8')

all_combinations['same_district'].sum()

all_combinations.drop(['u_postal', 'u_district', 'u_zone', 'g_postal', 'zone_id', 'district_id', 'postal_code'], axis=1, inplace=True)

all_combinations.head()

group_target_real = all_combinations.merge(
    group_target_real[['user_id', 'group_id', 'n_visits', 'visited']], 
    on=['user_id', 'group_id'], how='left').fillna(0)

group_target_real.info()

group_target_real['n_visits'] = group_target_real['n_visits'].astype('int8')
group_target_real['visited'] = group_target_real['visited'].astype('int8')

len(group_target_real['user_id'].unique())

del all_combinations

## Данные по пользователям ранее

visited_raw = pd.read_sql("""
    SELECT 
        user_id
        , category2_id
        , group_id
        , weekday
        , lower_bound
        , upper_bound
        , count(user_id) as n
        , is_online
    FROM (
        SELECT 
            attend.group_id
            , user_id
            , strftime('%w', event_date) + 1 as weekday
            , CAST(event_start AS INT) as lower_bound
            , CAST(event_end AS INT) as upper_bound
            , attend.is_online
            , category2_id
        FROM attend
        LEFT JOIN groups ON attend.group_id = groups.group_id
        WHERE event_date < '2023-02-01' AND event_date >= '2023-01-01' AND category2_id IS NOT NULL
    )
    GROUP BY user_id, category2_id, group_id, weekday, lower_bound, upper_bound, is_online
""", con=db, dtype={
    'user_id': 'int32', 'category2_id': 'int16', 'weekday': 'int8', 
    'lower_bound': 'int8', 'upper_bound': 'int8',
    'n': 'int8'
})

visited_raw['u_timetable'] = visited_raw.apply(
    lambda x: week_ones(x['weekday'], x['lower_bound'], x['upper_bound'], x['n'], padding=0)
, axis=1)

visited_raw['u_timetable_time'] = visited_raw.apply(
    lambda x: week_ones(x['weekday'], x['lower_bound'], x['upper_bound'], x['n'], padding=2).sum(axis=0)
, axis=1)

week_ones(1, 9, 12, 2, padding=2).sum(axis=0)

user_real_visits = visited_raw.groupby(['user_id', 'group_id'], as_index=False).agg({'n': 'sum'})
user_real_visits.columns = ['user_id', 'group_id', 'target_n_visited']

from scipy.special import softmax

user_visit_time_general = visited_raw.groupby('user_id', as_index=False).agg({'u_timetable_time': sum})
# user_visit_time_general['u_timetable_time'] = user_visit_time_general['u_timetable_time'].apply(lambda x: np.clip(x, 0, 1))
user_visit_time_general['u_timetable_time'] = user_visit_time_general['u_timetable_time'].apply(lambda x: (softmax(x) / max(softmax(x))).round(2))

user_visit_time = visited_raw.groupby('user_id', as_index=False).agg({'u_timetable': sum})

user_visit_time.loc[20]['u_timetable']

user_online = visited_raw.groupby('user_id', as_index=False).agg({'n': 'sum', 'is_online': 'sum'})

user_online['is_offline'] = user_online['n'] - user_online['is_online']

user_visit_time_cat = visited_raw.groupby(['user_id', 'category2_id'], as_index=False).agg({'u_timetable': sum})
user_visit_time_cat.columns = ['user_id', 'category2_id', 'u_cat_timetable']

user_online.head()

user_online.columns = ['user_id', 'n_visits', 'n_online', 'n_offline']

user_online[['n_online', 'n_offline']] = softmax(np.log(user_online[['n_online', 'n_offline']] + 1), axis=1)

user_online.head()

user_online = user_online[['user_id', 'n_online', 'n_offline']]
user_online.columns = ['user_id', 'u_online', 'u_offline']

total = group_target_real.merge(
    user_visit_time, on='user_id'
).merge(
    user_online, on='user_id'
).merge(
    user_real_visits, on=['group_id', 'user_id'], how='left'
).merge(
    user_visit_time_cat, on=['category2_id', 'user_id'], how='left'
).merge(
    user_visit_time_general, on='user_id'
)

total['n_visits'] = total['n_visits'].fillna(0).astype('int8')
total['n_neighbors'] = total['n_neighbors'].astype('int16')
total = total[total['n_visits'] == 0]

total = total.merge(
    total[total['target_n_visited'] > 0][['user_id', 'category3_id']].drop_duplicates()
)

total.shape

total['u_cat_timetable'] = total['u_cat_timetable'].apply(lambda x: np.zeros((7, 17), dtype='int8') if type(x) == float else x)

total['is_offline'] = 1 - total['is_online'] 

total['u_online'] = total['u_online'].astype('float16')
total['u_offline'] = total['u_offline'].astype('float16')

total['unmatch_other'] = total.progress_apply(
    lambda x: 1 - ((1 - np.clip((x['u_timetable'] - x['u_cat_timetable']) / 4, 0, 1) ) * x['timetable']).sum() / x['timetable'].sum(),
    axis=1
)

total['daytime_match'] = total.progress_apply(
    lambda x: (x['u_timetable_time'] * x['timetable']).sum() / x['timetable'].sum(),
    axis=1
)

total['match_same'] = total.progress_apply(
    lambda x: 1 - ((1 - np.clip(x['u_cat_timetable'] / 4, 0, 1)) * x['timetable']).sum() / x['timetable'].sum(),
    axis=1
)

total['target_n_visited'] = total['target_n_visited'].fillna(0).astype('int16')
total['unmatch_other'] = total['unmatch_other'].astype('float32')
total['match_same'] = total['match_same'].astype('float32')

(total[total['daytime_match'] == 0]['target_n_visited'] > 0).mean()

(total[total['daytime_match'] > 0]['target_n_visited'] > 0).mean()

total.drop(['u_timetable', 'u_cat_timetable', 'timetable', 'u_timetable_time'], axis=1, inplace=True)

total.columns

total.columns = ['group_id', 'n_days', 'category3_id', 'category2_id', 'g_online', 'g_mobile',
       'user_id', 'same_post', 'same_zone', 'same_district', 'n_neighbors',
       'n_visits', 'visited', 'u_online', 'u_offline',
       'target_n_visited', 'g_offline', 'unmatch_other', 'daytime_match', 'match_same']

total['x_online'] = (total['g_online'] * total['u_online']).round(4).astype('float16')
total['x_offline'] = (total['g_offline'] * total['u_offline']).round(4).astype('float16')

total.columns

total = total[[
        'group_id', 'user_id', 'category3_id',
        'x_online', 'x_offline', 
        'same_post', 'same_zone', 'same_district',
        'unmatch_other', 
        'match_same', 
        'daytime_match',
        'n_neighbors', 
        'n_days',
        'target_n_visited'
]]

total.head()

total['has_neighbors'] = (total['n_neighbors'] > 0).astype('int8')
total['log_neighbors'] = np.log(total['n_neighbors'] + 1).astype('float32')
total['log_days'] = np.log(total['n_days'] + 1).astype('float32')

cols = [
#        'x_online', 'x_offline', 
        'same_post', 'same_zone', 'same_district',
        'unmatch_other', 'match_same', 'daytime_match',
        'has_neighbors', 'log_days'
    ]

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report

users_train, users_test = train_test_split(total['user_id'].unique(), test_size=0.2)

i_train = total['user_id'].isin(set(users_train))
i_test = total['user_id'].isin(set(users_test))

X_train = total[i_train].set_index(['user_id', 'category3_id'])[cols]
X_test = total[i_test].set_index(['user_id', 'category3_id'])[cols]
y_train = (total[i_train].set_index(['user_id', 'category3_id'])['target_n_visited'] > 0).astype('int8')
y_test = (total[i_test].set_index(['user_id', 'category3_id'])['target_n_visited'] > 0).astype('int8')

model = LogisticRegression(class_weight='balanced')

model.fit(X_train, y_train)

print(classification_report(y_test, model.predict(X_test)))

print(model.intercept_)
for n, c in zip(X_train.columns, model.coef_[0]):
    print(c, n)

x_see = X_test.copy()
x_see['n_occurred'] = total[i_test]['target_n_visited'].values
x_see['has_occurred'] = (total[i_test]['target_n_visited'] > 0).astype('int8').values
x_see['group_id'] = total[i_test]['group_id'].values
x_see['predict'] = model.predict_proba(X_test)[:, 1].round(8)
x_see['predict'] = x_see['predict']#.astype('float16')
# x_see['has_neighbors'] = (x_see['n_neighbors'] > 0).astype('int8')
x_see = x_see.reset_index()

x_see = x_see.sort_values(
#     by='predict'
#     , ascending=False
    by=['predict', 'same_district', 'same_zone', 'has_neighbors', 'same_post', 'group_id'] #, 'match_same'
    , ascending=[False, False, False, False, False, False]
).groupby(
    ['user_id', 'category3_id'], as_index=False
).agg({
    'predict': list,
    'group_id': list,
    'has_occurred': list,
    'n_occurred': list,
    'same_post': list,
    'same_zone': list,
    'same_district': list,
    'has_neighbors': list,
#     'unmatch_other': list, 
#     'daytime_match': list, 
    'log_days': list, 
#     'match_same': list
})

x_see['category_real'] = x_see.apply(
    lambda x: [i[1] for i in sorted(zip(x['n_occurred'], x['group_id'])) if i[0] > 0],
    axis=1
)

x_see = x_see[x_see['category_real'].apply(len) > 0]

x_see['candidates'] = x_see.apply(
    lambda x: x['group_id'][
        :min((np.argmax(
            np.array(x['predict'][:-1]) - np.array(x['predict'][1:])
        ) if len(x['predict']) > 1 else 0) + 1, 10)
    ]
    , axis=1
)

x_see['share_uq'] = x_see.apply(
    lambda x: max([1 if real in x['group_id'][:1] else 0 for real in x['category_real']])
    , axis=1
)
x_see['share_uq'].mean()

x_see['idx'] = x_see.apply(
    lambda x: np.min([x['candidates'].index(real) if real in x['candidates'] else -1 for real in x['category_real'] ])
    , axis=1
)
x_see['mx_idx'] = x_see['candidates'].apply(len)
# x_see['share_uq'].mean()

# x_see[['mx']]

x_see['share_uq'] = x_see.apply(
    lambda x: np.max([1 if real in x['candidates'] else 0 for real in x['category_real']])
    , axis=1
)
x_see['share_uq'].mean()

x_see[x_see['idx'] > 1].sort_values(by='idx')[['predict', 'candidates', 'category_real', 'same_district', 'same_zone', 'same_post', 'log_days', 'idx', 'mx_idx']]

x_see[x_see['idx'] == 1].sort_values(by='mx_idx')#.head(50)

plt.figure(figsize=(20, 10))
sns.scatterplot(x='mx_idx', y='idx', data=x_see, alpha=0.02)
# plt.xlim((-2, 11))
# plt.ylim((-2, 10))
plt.show()

# x_see['mx_idx'].hist(bins=100)

# plt.figure(figsize=(20, 10))
# sns.scatterplot(x='mx_idx', y='idx', data=x_see, alpha=0.02)
# plt.show()
x_see[x_see['idx'] < 50]['idx'].hist(bins=100)

group_target[group_target['group_id'].isin({801349222, 801366063})]

# x_see[x_see['share_uq'] > 0]

# x_see.info(memory_usage = "deep")
